#20 Pet registry
pet views separate from main
