+++ app/frontend/templates/base.html
@@ -12,7 +12,7 @@
                     </li>
 
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for("main.pets") }}">Pets</a>
+                        <a class="nav-link" href="{{ url_for("pet.pets") }}">Pets</a>
                     </li>
                 </ul>
