+++ app/frontend/templates/pet-add.html
@@ -6,7 +6,7 @@
 {% endblock %}
 
 {% block content %}
-    <form action="{{ url_for("main.pet_add") }}" method="POST">
+    <form action="{{ url_for("pet.pet_add") }}" method="POST">
         <div class="form-control">
             <label for="name">Name</label>
             <input type="text" class="form-control" name="name" id="name">
